Naruto/requirements.txt
request==2.24.0
bs4==0.0.1
beautifulsoup4==4.9.3
Pillow==8.0.0
PyPDF2==1.26.0